+++ src/types/FabloConfigExtended.ts
@@ -86,6 +86,18 @@ export interface RootOrgConfig {
   ordererHead: OrdererConfig;
 }
 
+export interface FabloRestConfig {
+  address: string;
+  port: number;
+  affiliation: string;
+  mspId: string;
+  fabricCaUrl: string;
+  fabricCaName: string;
+  discoveryUrls: string;
+  discoverySslTargetNameOverrides: string;
+  discoveryTlsCaCertFiles: string;
+}
+
 export interface OrgConfig {
   anchorPeers: PeerConfig[];
   bootstrapPeers: string;
@@ -98,6 +110,7 @@ export interface OrgConfig {
   name: string;
   peers: PeerConfig[];
   peersCount: number;
+  tools: { fabloRest: FabloRestConfig | undefined };
 }
 
 export interface ChaincodeConfig {