+++ e2e-network/test-03-private-data.sh
@@ -34,7 +34,7 @@ waitForChaincode() {
 }
 
 expectInvoke() {
-  sh "$TEST_TMP/../expect-invoke.sh" "$1" "$2" "$3" "$4" "$5" "$6" "$7"
+  sh "$TEST_TMP/../expect-invoke-cli.sh" "$1" "$2" "$3" "$4" "$5" "$6" "$7"
 }
 
 trap networkDown EXIT
@@ -54,6 +54,8 @@ waitForChaincode "cli.org1.com" "peer0.org1.com:7060" "my-channel1" "or-policy-c
 waitForChaincode "cli.org2.com" "peer0.org2.com:7070" "my-channel1" "or-policy-chaincode" "0.0.1"
 waitForChaincode "cli.org1.com" "peer0.org1.com:7060" "my-channel1" "and-policy-chaincode" "0.0.1"
 waitForChaincode "cli.org2.com" "peer0.org2.com:7070" "my-channel1" "and-policy-chaincode" "0.0.1"
+
+sleep 3 # extra time needed: peers need to discover themselves before private data call.
 
 # Org1: Test chaincode with transient fields and private data
 expectInvoke "cli.org1.com" "peer0.org1.com:7060" "my-channel1" "or-policy-chaincode" \