+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## 0.3.0-unstable
 
 ### Features
-
+* Add [Fablo REST](https://github.com/softwaremill/fablo-rest/) support 
 * By default are peers are anchor peers
 
 ### Fixes
+* Fixed issue with bad requiredPeerCount in private data collection
 
 ### Chore & Maintenance