+++ CHANGELOG.md
@@ -1,44 +1,61 @@
 ## 0.3.0-unstable
+
 ### Features
+* Add [Fablo REST](https://github.com/softwaremill/fablo-rest/) support 
+* By default all peers are anchor peers
+
 ### Fixes
+* Fixed issue with bad requiredPeerCount in private data collection
+
 ### Chore & Maintenance
 
 ## 0.2.0
+
 * Rename Fabrica to Fablo
 
 ## 0.1.1
+
 * Broken Node.js chaincode build ([#211](https://github.com/softwaremill/fablo/pull/211))
 
 ## 0.1.0
 
 ### Features
+
 * Support for private data ([#104](https://github.com/softwaremill/fablo/issues/104))
 * Added channel query scripts  ([#169](https://github.com/softwaremill/fablo/issues/169))
-* Support for Hyperledger Fabric 2.x ([#132](https://github.com/softwaremill/fablo/issues/132), [#178](https://github.com/softwaremill/fablo/issues/178), [#190](https://github.com/softwaremill/fablo/issues/178))
+* Support for Hyperledger Fabric 2.x ([#132](https://github.com/softwaremill/fablo/issues/132)
+  , [#178](https://github.com/softwaremill/fablo/issues/178), [#190](https://github.com/softwaremill/fablo/issues/178))
 * Support default endorsement policy ([#189](https://github.com/softwaremill/fablo/issues/189))
 * Support for fablo config in YAML format
 
 ### Chore & Maintenance
+
 * Use different config format and provide defaults
 * Rewrite Yeoman generators to use TypeScript
 
 ## 0.0.1
 
 ### Features
+
 * Generate simple Hyperledger Fabric network
 * Support for multiple organizations
 * Support for solo consensus protocol
-* Support for RAFT consensus protocol ([#16](https://github.com/softwaremill/fablo/issues/16), [#38](https://github.com/softwaremill/fablo/issues/38))
+* Support for RAFT consensus protocol ([#16](https://github.com/softwaremill/fablo/issues/16)
+  , [#38](https://github.com/softwaremill/fablo/issues/38))
 * Validation of `fablo-config.json` based on JSON schema and other rules
 * Allow to upgrade chaincode ([#45](https://github.com/softwaremill/fablo/issues/36))
 * Network recreation in one step ([#105](https://github.com/softwaremill/fablo/issues/105))
-* Init command to provide simple config ([#90](https://github.com/softwaremill/fablo/issues/90)) with chaincode ([#100](https://github.com/softwaremill/fablo/issues/100))
-* Proper exposing Orderer and Peers ports for service discovery ([#116](https://github.com/softwaremill/fablo/issues/116))
+* Init command to provide simple config ([#90](https://github.com/softwaremill/fablo/issues/90)) with
+  chaincode ([#100](https://github.com/softwaremill/fablo/issues/100))
+* Proper exposing Orderer and Peers ports for service
+  discovery ([#116](https://github.com/softwaremill/fablo/issues/116))
 
 ### Fixes
+
 * Missing notify anchor peers step ([#26](https://github.com/softwaremill/fablo/issues/26))
 
 ### Chore & Maintenance
+
 * Lint bash and YAML files ([#48](https://github.com/softwaremill/fablo/issues/48))
 * Format generated scripts and YAML files ([#75](https://github.com/softwaremill/fablo/issues/75))
 * Test Java chaincode on generated network ([#25](https://github.com/softwaremill/fablo/issues/25))